Fix some minor issues
