+++ frontend/lib/about/dev_settings_view.dart
@@ -1,10 +1,10 @@
 import 'dart:convert';
 import 'dart:developer';
-import 'package:base32/base32.dart';
 
+import 'package:base32/base32.dart';
 import 'package:ehrenamtskarte/configuration/settings_model.dart';
+import 'package:ehrenamtskarte/identification/activation/activation_code_parser.dart';
 import 'package:ehrenamtskarte/identification/activation_code_model.dart';
-import 'package:ehrenamtskarte/identification/activation_code_parser.dart';
 import 'package:ehrenamtskarte/identification/qr_code_scanner/qr_code_processor.dart';
 import 'package:ehrenamtskarte/intro_slides/intro_screen.dart';
 import 'package:ehrenamtskarte/proto/card.pb.dart';
