--- frontend/lib/identification/card/id_card.dart
@@ -1,27 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// Manually chosen aspect ratio of a card. Maybe this was determined using
-/// a ruler?
-const creditCardAspectRatio = 86 / 54;
-
-class IdCard extends StatelessWidget {
-  final Widget child;
-
-  const IdCard({super.key, required this.child});
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      elevation: 5,
-      color: Colors.white,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(10.0),
-      ),
-      clipBehavior: Clip.antiAlias,
-      child: ConstrainedBox(
-        constraints: const BoxConstraints(maxHeight: 600, maxWidth: 600),
-        child: AspectRatio(aspectRatio: creditCardAspectRatio, child: child),
-      ),
-    );
-  }
-}
